requirement.txt
absl-py==0.13.0
addict==2.4.0
albumentations==1.1.0
argon2-cffi==21.1.0
attrs==21.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
chumpy==0.70
click==8.0.1
colorama==0.4.4
coverage==6.0.1
cycler==0.10.0
Cython==0.29.24
debugpy @ file:///tmp/build/80754af9/debugpy_1629222698064/work
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml==0.7.1
entrypoints==0.3
filelock==3.0.12
flake8==3.9.2
flake8-import-order==0.18.1
flatbuffers==2.0
future==0.18.2
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.23
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
idna==3.2
imageio==2.9.0
iniconfig==1.1.1
interrogate==1.5.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1629995100487/work/dist/ipykernel-6.2.0-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1632141392005/work
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.21
jedi @ file:///tmp/build/80754af9/jedi_1611333125342/work
Jinja2==3.0.1
joblib==1.1.0
json-tricks==3.15.5
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1630690655192/work
jupyter-console==6.4.0
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213311222/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mmcv-full==1.3.14
mmpose==0.19.0
motmetrics==1.2.0
munkres==1.1.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.6.3
notebook==6.4.4
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626271506491/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
onnx==1.10.1
onnxruntime==1.9.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1625655817137/work
Pillow-SIMD==7.0.0.post3
pluggy==1.0.0
poseval @ git+https://github.com/svenkreiss/poseval.git@3128c5cbcf90946e5164ff438ad651e113e64613
progressbar2==3.53.2
prometheus-client==0.11.0
prompt-toolkit @ file:///Users/ktietz/demo/mc3/conda-bld/prompt-toolkit_1629146841569/work
protobuf==3.17.3
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py==1.10.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
pytest-benchmark==3.4.1
pytest-runner==5.3.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-utils==2.5.6
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq @ file:///tmp/build/80754af9/pyzmq_1628276010766/work
qtconsole==5.1.1
QtPy==1.11.2
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
Send2Trash==1.8.0
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==4.0.0
smplx==0.1.28
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
thinplate @ git+https://github.com/cheind/py-thin-plate-spline@f6995795397118b7d0ac01aecd3f39ffbfad9dee
threadpoolctl==3.0.0
tifffile==2021.8.30
toml==0.10.2
torch==1.8.2
torch-tb-profiler==0.2.1
torchaudio==0.8.2
torchsummary==1.5.1
torchvision==0.9.2
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.62.2
traitlets @ file:///tmp/build/80754af9/traitlets_1632522747050/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.6
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
xdoctest==0.15.10
xmltodict==0.12.0
xtcocotools==1.10
yapf==0.31.0
